+++ include/hm_map.h
@@ -56,8 +56,9 @@ typedef struct hm_map {
     size_t size;
     size_t len;
 
-    hm_cmp cmp;
-    hm_hash hash;
+    hm_cmp cmp_key;
+    hm_hash hash_key;
+    
     hm_free free_key;
     hm_free free_val;
 } hm_map;